Create MainActivity.java
Main Activity Page:
It is the opening page of the application with Title and Logo of the brand. This gives user a good appeal about the application and the event management organization that they would look forward to working with.
Designed using Text View for giving a suitable Title. Added the logo using image design, added a button that takes you to the next activity pageà’About Page’