Event Description Page
This page has the details of the event described along with the service description. There is an image included giving a basic idea of how the event would look when they decorate and make arrangements. The user is asked three questions for availing their services.
The questions include an Editable Text Field (input type-numbers), a radio button group with ‘yes’ or ‘no’ radio buttons and a spinner with three options to select.