Feedback Page
This page had an email design and a send button. I used implicit intents to send email. An intent object is used with (ACTION_SEND) and data to launch the available email clients to send an email from our application.
This application gives a full-fledged satisfaction for our customers to use our services.